refactor(mesh): migrate to u32 indices and extract mesh processing
- Change mesh indices from u16 to u32 to support larger meshes
- Extract marching tetrahedra algorithm to separate module
- Add tissue-based mesh generation with material properties
- Improve vertex normal calculation and smoothing
- Update render pass to use Uint32 index format

+++ src/rendering/view/mesh/basic_mesh_context.rs
@@ -183,7 +183,7 @@ impl BasicMeshContext {
         render_pass.set_bind_group(0, &self.bind_group, &[]);
         render_pass.set_bind_group(1, &self.lighting_bind_group, &[]);
         render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
-        render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
+        render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint32);
         render_pass.draw_indexed(0..self.num_indices, 0, 0..1);
 
         log::trace!("BasicMeshContext::render - Draw call completed");
